refactor(gophermart): rename commonErrors to customErrors
