+++ internal/gophermart/customErrors/customErrors.go
@@ -1,4 +1,4 @@
-package errors
+package customErrors
 
 import "errors"
