+++ internal/gophermart/handlers/auth.go
@@ -3,7 +3,7 @@ package handlers
 import (
 	"errors"
 	"net/http"
-	commonErrors "ya41-56/internal/gophermart/errors"
+	"ya41-56/internal/gophermart/customErrors"
 	"ya41-56/internal/gophermart/models"
 	"ya41-56/internal/gophermart/services"
 	"ya41-56/internal/shared/contextutil"
@@ -58,7 +58,7 @@ func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
 
 	jwtToken, err := h.Auth.Login(r.Context(), req.Login, req.Password)
 	if err != nil {
-		if errors.Is(err, commonErrors.ErrInvalidCreds) {
+		if errors.Is(err, customErrors.ErrInvalidCreds) {
 			response.Error(w, http.StatusUnauthorized, err.Error())
 		} else {
 			response.Error(w, http.StatusInternalServerError, err.Error())
@@ -99,7 +99,7 @@ func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
 		Password: req.Password,
 	})
 	if err != nil {
-		if errors.Is(err, commonErrors.ErrUserExists) {
+		if errors.Is(err, customErrors.ErrUserExists) {
 			response.Error(w, http.StatusConflict, err.Error())
 		} else {
 			response.Error(w, http.StatusInternalServerError, err.Error())
