+++ internal/gophermart/services/auth.go
@@ -2,7 +2,7 @@ package services
 
 import (
 	"context"
-	commonErrors "ya41-56/internal/gophermart/errors"
+	"ya41-56/internal/gophermart/customErrors"
 	"ya41-56/internal/gophermart/models"
 	"ya41-56/internal/shared/bcryptutil"
 	"ya41-56/internal/shared/repositories"
@@ -23,15 +23,15 @@ func NewAuthService(users repositories.Repository[models.User], tokenService *To
 func (s *AuthService) Login(ctx context.Context, login, password string) (string, error) {
 	user, err := s.Users.FindByField(ctx, "login", login)
 	if err != nil {
-		return "", commonErrors.ErrInvalidCreds
+		return "", customErrors.ErrInvalidCreds
 	}
 
 	if err = bcryptutil.CheckHash(password, user.PasswordHash); err != nil {
-		return "", commonErrors.ErrInvalidCreds
+		return "", customErrors.ErrInvalidCreds
 	}
 
 	if user.Status == models.UserStatusDisabled {
-		return "", commonErrors.ErrInvalidCreds
+		return "", customErrors.ErrInvalidCreds
 	}
 
 	return s.TokenService.BuildJWTString(user.ID)
@@ -41,7 +41,7 @@ func (s *AuthService) ParseAndValidate(tokenString string) (string, error) {
 	claims := Claims{}
 	token, err := s.TokenService.ParseToken(&claims, tokenString)
 	if err != nil || !token.Valid {
-		return "", commonErrors.ErrJWTToken
+		return "", customErrors.ErrJWTToken
 	}
 
 	return claims.UserID, nil
@@ -50,7 +50,7 @@ func (s *AuthService) ParseAndValidate(tokenString string) (string, error) {
 func (s *AuthService) Register(ctx context.Context, user *models.User) (string, error) {
 	_, err := s.Users.FindByField(ctx, "login", user.Login)
 	if err == nil {
-		return "", commonErrors.ErrUserExists
+		return "", customErrors.ErrUserExists
 	}
 
 	hashed, err := bcryptutil.Hash(user.Password)
