+++ internal/gophermart/services/base_test.go
